backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
pymongo==4.6.3
celery==5.3.6
redis==5.0.3
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.5
eventlet==0.35.2
flask-socketio==5.3.6
flask-limiter==3.7.0
psutil==5.9.5
prometheus-flask-exporter==0.23.0
prometheus-client==0.19.0
openai==1.14.0
google-generativeai==0.4.1
marshmallow==3.21.1
docker==7.0.0
